LayoutTests/imported/w3c/web-platform-tests/resource-timing/TAO-match-expected.txt


FAIL The timing allow check algorithm will pass when the Timing-Allow-Origin header value list contains a case-sensitive match. assert_greater_than: transferSize should be greater than 0 expected a number greater than 0 but got 0
FAIL The timing allow check algorithm will pass when the Timing-Allow-Origin header value list contains a wildcard. assert_greater_than: transferSize should be greater than 0 expected a number greater than 0 but got 0
PASS The timing allow check algorithm will fail when the Timing-Allow-Origin header value list contains a null origin.
FAIL The timing allow check algorithm will pass when the Timing-Allow-Origin header value list contains the origin and a wildcard. assert_greater_than: transferSize should be greater than 0 expected a number greater than 0 but got 0
FAIL The timing allow check algorithm will pass when the Timing-Allow-Origin header value list contains multiple wildcards. assert_greater_than: transferSize should be greater than 0 expected a number greater than 0 but got 0
FAIL The timing allow check algorithm will pass when the Timing-Allow-Origin header value contains only the origin. assert_greater_than: transferSize should be greater than 0 expected a number greater than 0 but got 0
PASS The timing allow check algorithm will fail when the Timing-Allow-Origin header value contains only the uppercased origin.
PASS The timing allow check algorithm will fail when the Timing-Allow-Origin header value contains only a space.
FAIL The timing allow check algorithm will fail when the Timing-Allow-Origin header value contains only a wildcard. assert_greater_than: transferSize should be greater than 0 expected a number greater than 0 but got 0
PASS The timing allow check algorithm will fail when the Timing-Allow-Origin header is not present.
